Waad changes
